Finished Collectible
Finished the collectible turorial from Danny's in class.
Applied it to an BP_Intel, but used the coin mesh that was provided. Will change to an intel mesh that I create.
I want to also change the rotation of the coin, to an up and down motion for the intel, and then make it move up and down quickly once collected.